requirements.txt
atlassian-python-api==3.12.0
certifi==2021.5.30
chardet==4.0.0
Deprecated==1.2.12
idna==2.10
oauthlib==3.1.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.5
wrapt==1.12.1
